Cleanup the bakery cukes and add a delete flash
The cukes had a few steps that were `expect(page).to have_content(...)` without any kind of filter (e.g. `within()`). It also suffered from the "should be redirected" not being a test but actually navigating you. Lastly, `When` is a user interaction, rarely will that be typing a url, `given` that you are on a page fits a bit better.

Ensuring you are on a page is less valuable then the info or reason you are at that page. e.g. info about the bakery vs the `bakery_edit path`