Speedup prawn using bakery logos
Prawn was copying the logo from s3 for each invoice, this was slow. This method (if necessary) copies the logo once and saves the temporary file on an instance variable. The file will exist as long as the bakery object exists, which hopefully is only as long as the request is active.

I’ve run these tests with and without s3 keys