Speed up the ShipmentCreator
Increased creating shipments from 6-10 a minute to 3-5 hundred a minute. We could still do better as we’re doing a lot of repeated work around lead times. The hidden area of work is when we save a shipment item and it asks the product it’s lead time, no caching there. The active record query cache helps a ton but still we could do better. Redis or memcache? A counter cache? Cache invalidation is hard.

We now make shipments for every client in advance as the max_lead_time for all products never mind ordered products. That sucks a bit but is performant.

 - Laxup the client validations on shipments
 - Remove the skylight stuff
 - cleanup the KickoffService ShipmentService and ProductionRunService specs
 - Would be lovely if we could keep a cache of recipe.total_lead_time