Stop calling debounce so much.
On some browsers window.onResize gets called a *lot*. The debounce function returns a function that is safe to call a lot. It will wait 250ms before calling the callback it received and ignore repeated calls during that time. The way we had it created a new debounced function each and every time, each one waiting 250ms to call the `resetCenter` and then getting thrown away. We now save the debounced `resetCenter` and use it over and over.