Improve `OrderItem.production_date` by 89%
- Complexity of the old query was 22212.20
- Complexity of the new query is 2500

Thank you stack exchange!
http://dba.stackexchange.com/questions/108623/running-a-complex-query-for-every-date-in-a-range
http://dba.stackexchange.com/questions/108767/a-costly-approach-to-filtering-on-a-calculated-date

They suggested the `select distinct` which is a lot easier to understand. Since it became a lot easier to figure out what was what, I applied it to the order items as a whole instead of just orders. I also used an CTE which makes the sql read a lot easier and for some reason I have no idea perform better!?

I still believe this could be done a lot faster but it's fast enough for now!